+++ docs/changes.md
@@ -1,4 +1,8 @@
 # 最新动态
+2022/06/07
+  * combo box 增加theme\_of\_popup属性。
+  * 增加文档 [如何设置下拉选择框 (combo box) 弹出窗口/列表项的样式](how_to_set_style_of_combo_box_item.md)
+
 2022/06/04
   * 增加函数 conf\_doc\_set\_node\_prop